tests/validation/requirements.txt
git+https://github.com/rancher/client-python.git
boto3
cryptography
flake8
invoke
Jinja2
netaddr
paramiko
PyJWT
pytest-repeat
pytest-xdist
pytest
pytest-html
pytest-ordering
python-digitalocean
PyYAML
requests
websocket-client
pytest-lazy-fixture
semver
pyproject-toml
python-ldap
urllib3
lxml
docker-image-py
kubernetes
jsonpath-ng
mysql-connector-python
